@@ -1,16 +1,16 @@
 {
   "dependencies": {
-    "@angular/animations": "^5.0.3",
-    "@angular/cdk": "^5.0.3",
-    "@angular/common": "^5.0.3",
-    "@angular/compiler": "^5.0.3",
-    "@angular/core": "^5.0.3",
-    "@angular/router": "^5.0.3",
-    "@angular/flex-layout": "5.0.0-beta.15",
-    "@angular/forms": "^5.0.3",
-    "@angular/material": "^5.0.3",
-    "@angular/platform-browser": "^5.0.3",
-    "@angular/platform-browser-dynamic": "^5.0.3",
+    "@angular/animations": "^6.0.7",
+    "@angular/cdk": "^6.0.7",
+    "@angular/common": "^6.0.7",
+    "@angular/compiler": "^6.0.7",
+    "@angular/core": "^6.0.7",
+    "@angular/router": "^6.0.7",
+    "@angular/flex-layout": "6.0.0-beta.16",
+    "@angular/forms": "^6.0.7",
+    "@angular/material": "^6.0.7",
+    "@angular/platform-browser": "^6.0.7",
+    "@angular/platform-browser-dynamic": "^6.0.7",
     "@types/es6-promise": "0.0.33",
     "@types/jasmine": "^2.8.2",
     "@types/node": "^8.0.53",
@@ -18,7 +18,7 @@
     "html-webpack-plugin": "^2.30.1",
     "reflect-metadata": "^0.1.10",
     "roboto-fontface": "^0.7.0",
-    "rxjs": "^5.0.0-beta.6",
+    "rxjs": "^6.0.7",
     "zone.js": "^0.8.18"
   },
   "scripts": {
